requirements.txt
pandas=1.2.4
pyyaml=3.12
pydantic=1.8.2
streamlit=0.82.0
pytest=6.2.4